Remove support for the BCM mappings
This architecture specific mapping is a red herring for users.

The pins used for the BCM mappings vary between different RPi
versions, so arbiratrily making the Orange Pi XXX look like
a Raspberry Pi does not make sense.

Further, the BCM mappings for the Orange Pi PC (and all the
boards that share it's pinouts) are just a copy of the BOARD
layout, so they don't align with the intent of the mapping.

It's best to remove support for the BCM mapping and force an
error to the user (so they know they need to choose a different
mapping), rather than pretend to be compatible, but hand out
incorrect or irrelevant mappings.

Signed-off-by: Alastair D'Silva <alastair@d-silva.org>